requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
opencv-python==4.8.1.78
ultralytics==8.1.0
ocrmac==0.1.0
Pillow==10.1.0
numpy==1.24.3


